requirements.txt
django
djangorestframework
openai
requests
django-redis
decouple
psycopg2
eventlet
django-ratelimit
